[AspNetCore] Splits PublishToFolder Command
