+++ main/src/addins/MonoDevelop.AspNetCore/MonoDevelop.AspNetCore.csproj
@@ -37,11 +37,13 @@
     <Compile Include="MonoDevelop.AspNetCore\AspNetCoreCertificateManager.cs" />
     <Compile Include="MonoDevelop.AspNetCore\DotNetCoreDevCertsTool.cs" />
     <Compile Include="MonoDevelop.AspNetCore\CertificateCheckResult.cs" />
-    <Compile Include="MonoDevelop.AspNetCore.Commands\PublishCommandHandler.cs" />
     <Compile Include="MonoDevelop.AspNetCore.Commands\PublishCommandItem.cs" />
     <Compile Include="MonoDevelop.AspNetCore.Commands\ProjectPublishProfile.cs" />
     <Compile Include="MonoDevelop.AspNetCore.Dialogs\PublishToFolderDialog.cs" />
     <Compile Include="MonoDevelop.AspNetCore\ProjectExtensions.cs" />
+    <Compile Include="MonoDevelop.AspNetCore.Commands\PublishToFolderCommandHandler.cs" />
+    <Compile Include="MonoDevelop.AspNetCore.Commands\PublishToFolderProfilesCommandHandler.cs" />
+    <Compile Include="MonoDevelop.AspNetCore.Commands\PublishToFolderBaseCommandHandler.cs" />
   </ItemGroup>
   <ItemGroup>
     <EmbeddedResource Include="Properties\MonoDevelop.AspNetCore.addin.xml" />
