+++ main/src/addins/MonoDevelop.AspNetCore/Properties/MonoDevelop.AspNetCore.addin.xml
@@ -419,25 +419,29 @@
 	</ExtensionPoint>
 	
 	<Extension path = "/MonoDevelop/Ide/Commands/Project">
-		<Command id = "MonoDevelop.AspNetCore.Commands.PublishToFolder" 
-				defaultHandler = "MonoDevelop.AspNetCore.Commands.PublishCommandHandler"
+		<Command id = "MonoDevelop.AspNetCore.Commands.PublishToFolderProfiles" 
+				defaultHandler = "MonoDevelop.AspNetCore.Commands.PublishToFolderProfilesCommandHandler"
 				type="array"
+				_label = "Publish" />
+		<Command id = "MonoDevelop.AspNetCore.Commands.PublishToFolder" 
+				defaultHandler = "MonoDevelop.AspNetCore.Commands.PublishToFolderCommandHandler"
 				_label = "Publish" />
 	</Extension>
 	
 	<Extension path = "/MonoDevelop/Ide/ContextMenu/ProjectPad">		
 		<Condition id="ItemType" value="Project">
-			<ItemSet id = "PublishToFolder" _label = "Publish" autohide="True" insertafter="MonoDevelop.Ide.ArchiveCommands.ViewArchives" >
+			<ItemSet id = "PublishToFolderProfiles" _label = "Publish" autohide="True" insertafter="MonoDevelop.Ide.ArchiveCommands.ViewArchives" >
 				<CommandItem id = "MonoDevelop.AspNetCore.Commands.PublishToFolder" />
+				<CommandItem id = "MonoDevelop.AspNetCore.Commands.PublishToFolderProfiles" />
 			</ItemSet>
 		</Condition>
 	</Extension>
 	
 	<Extension path = "/MonoDevelop/Ide/MainMenu/Build">		
 		<ItemSet id = "PublishToFolder" _label = "Publish" autohide="True" insertafter="MonoDevelop.Ide.ArchiveCommands.ViewArchives" >
-				<CommandItem id = "MonoDevelop.AspNetCore.Commands.PublishToFolder" />
+			<CommandItem id = "MonoDevelop.AspNetCore.Commands.PublishToFolder" />	
+			<CommandItem id = "MonoDevelop.AspNetCore.Commands.PublishToFolderProfiles" />
 		</ItemSet>
 	</Extension>
-
 	
 </ExtensionModel>
